Collect more stripe utilities; refactor collate_superstripe
